+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'freedombase:legal-management',
-  version: '1.2.0',
+  version: '1.3.0',
   summary: 'Manage your legal documents and user consent.',
   git: 'https://github.com/freedombase/meteor-legal-management',
   documentation: 'README.md'